requirements.txt
notion-client==1.0.0
python-dotenv==1.0.1
ruff
stravaio==0.0.9
